Added back possibility to use offline data

+++ src/app/impact/services/impact/impact-rest.service.ts
@@ -1,9 +1,10 @@
-import { ImpactDto } from '../../dtos/ImpactDto';
-import { RestSettings } from '../../settings/RestSettings';
-import { LogService } from '../../../shared/services/log.service';
-import { Observable } from 'rxjs';
-import { HttpClient } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import {ImpactDto} from '../../dtos/ImpactDto';
+import {RestSettings} from '../../settings/RestSettings';
+import {LogService} from '../../../shared/services/log.service';
+import {Observable, of} from 'rxjs';
+import {HttpClient} from '@angular/common/http';
+import {Injectable} from '@angular/core';
+import {SampleDataService} from "../../spec/sample-data.service";
 
 @Injectable({
   providedIn: 'root'
@@ -12,7 +13,8 @@ export class ImpactRestService {
 
   constructor(
     private logger: LogService,
-    private http: HttpClient) {
+    private http: HttpClient,
+    protected data: SampleDataService) {
 
   }
 
@@ -22,21 +24,37 @@ export class ImpactRestService {
 
   getImpactsByAnalysisId(analysisId: string): Observable<ImpactDto[]> {
     this.logger.info(this, 'Get Impact by AnalysisId');
-    return this.http.get<ImpactDto[]>(RestSettings.impactsUrl + "?analysisId=" + analysisId);
+    if (this.data.offline) {
+      return of(this.data.dummyImpactDtos);
+    } else {
+      return this.http.get<ImpactDto[]>(RestSettings.impactsUrl + "?analysisId=" + analysisId);
+    }
   }
 
   createImpact(impactDto: ImpactDto): Observable<any> {
     this.logger.info(this, 'Create Impact');
-    return this.http.post(RestSettings.impactsUrl, impactDto, RestSettings.httpOptions);
+    if (this.data.offline) {
+      return of(impactDto);
+    } else {
+      return this.http.post(RestSettings.impactsUrl, impactDto, RestSettings.httpOptions);
+    }
   }
 
   updateImpact(impactDto: ImpactDto): Observable<any> {
     this.logger.info(this, 'Update Impact');
-    return this.http.put(RestSettings.impactsUrl, impactDto, RestSettings.httpOptions);
+    if (this.data.offline) {
+      return of(impactDto);
+    } else {
+      return this.http.put(RestSettings.impactsUrl, impactDto, RestSettings.httpOptions);
+    }
   }
 
   deleteImpact(impactDto: ImpactDto): Observable<any> {
     this.logger.info(this, 'Delete Impact');
-    return this.http.delete(RestSettings.impactsUrl + '/' + impactDto.id);
+    if (this.data.offline) {
+      return of(impactDto);
+    } else {
+      return this.http.delete(RestSettings.impactsUrl + '/' + impactDto.id);
+    }
   }
 }
