+++ src/app/impact/services/value/value-rest.service.ts
@@ -3,8 +3,9 @@ import {ValueDto} from '../../dtos/ValueDto';
 import {LogService} from '../../../shared/services/log.service';
 import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
-import {Observable} from 'rxjs';
+import {Observable, of} from 'rxjs';
 import {ImpactDto} from "../../dtos/ImpactDto";
+import {SampleDataService} from "../../spec/sample-data.service";
 
 @Injectable({
   providedIn: 'root'
@@ -13,7 +14,8 @@ export class ValueRestService {
 
   constructor(
     private logger: LogService,
-    private http: HttpClient) {
+    private http: HttpClient,
+    protected data: SampleDataService) {
 
   }
 
@@ -23,31 +25,55 @@ export class ValueRestService {
 
   getValues(): Observable<ValueDto[]> {
     this.logger.info(this, 'Get all Values');
-    return this.http.get<ValueDto[]>(RestSettings.valuesUrl);
+    if (this.data.offline) {
+      return of(this.data.dummyValueDtos)
+    } else {
+      return this.http.get<ValueDto[]>(RestSettings.valuesUrl);
+    }
   }
 
   getValuesByAnalysisId(analysisId: string): Observable<ValueDto[]> {
     this.logger.info(this, 'Get all Values');
-    return this.http.get<ValueDto[]>(RestSettings.valuesUrl + "?analysisId=" + analysisId);
+    if (this.data.offline) {
+      return of(this.data.dummyValueDtos)
+    } else {
+      return this.http.get<ValueDto[]>(RestSettings.valuesUrl + "?analysisId=" + analysisId);
+    }
   }
 
   createValue(valueDto: ValueDto): Observable<any> {
     this.logger.info(this, 'Create Value');
-    return this.http.post(RestSettings.valuesUrl, valueDto, RestSettings.httpOptions);
+    if (this.data.offline) {
+      return of(valueDto);
+    } else {
+      return this.http.post(RestSettings.valuesUrl, valueDto, RestSettings.httpOptions);
+    }
   }
 
   updateValue(valueDto: ValueDto): Observable<any> {
     this.logger.info(this, 'Update Value');
-    return this.http.put(RestSettings.valuesUrl, valueDto, RestSettings.httpOptions);
+    if (this.data.offline) {
+      return of(valueDto);
+    } else {
+      return this.http.put(RestSettings.valuesUrl, valueDto, RestSettings.httpOptions);
+    }
   }
 
   deleteValue(valueDto: ValueDto): Observable<any> {
     this.logger.info(this, 'Delete Value');
-    return this.http.delete(RestSettings.valuesUrl + '/' + valueDto.id);
+    if (this.data.offline) {
+      return of(valueDto);
+    } else {
+      return this.http.delete(RestSettings.valuesUrl + '/' + valueDto.id);
+    }
   }
 
   getValueTypes(): Observable<string[]> {
     this.logger.info(this, 'Get all ValueTypes');
-    return this.http.get<string[]>(RestSettings.valueTypesUrl);
+    if (this.data.offline) {
+      return of(this.data.dummyValueTypes)
+    } else {
+      return this.http.get<string[]>(RestSettings.valueTypesUrl);
+    }
   }
 }
